Hide FileController implementation, create interface
